html/semantics/embedded-content-0/the-canvas-element/size.attributes.parse.octal.html

<!DOCTYPE html>
<title>Canvas test: size.attributes.parse.octal</title>
<meta name="author" content="Philip Taylor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../../../common/canvas-tests.js"></script>
<link rel="stylesheet" href="../../../../../common/canvas-tests.css">
<body class="show_output">

<h1><a href="index.size.html">size</a>.<a href="index.size.attributes.html">attributes</a>.<a href="index.size.attributes.parse.html">parse</a>.octal</h1>
<p class="desc">Parsing of non-negative integers</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.size.nonnegativeinteger">size.nonnegativeinteger</a>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="0100" height="0100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.parse.octal.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Parsing of non-negative integers");
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 100, "canvas.width", "100");
_assertSame(canvas.height, 100, "canvas.height", "100");
_assertEqual(window.getComputedStyle(canvas, null).getPropertyValue("width"), "100px", "window.getComputedStyle(canvas, null).getPropertyValue(\"width\")", "\"100px\"");
_assertSame(canvas.getAttribute('width'), '0100', "canvas.getAttribute('width')", "'0100'");
_assertSame(canvas.getAttribute('height'), '0100', "canvas.getAttribute('height')", "'0100'");


});
</script>
